(refacto) opening and closing
