+++ CMakeLists.txt
@@ -49,7 +49,7 @@ target_sources(threshold_cpu PRIVATE src/detect_obj.hpp src/threshold_cpu.cpp)
 add_library(detector_cpu)
 target_compile_features(detector_cpu PUBLIC cxx_std_14)
 target_link_libraries(detector_cpu PRIVATE spdlog::spdlog utilsimage blurring_cpu openclose_cpu threshold_cpu)
-target_sources(detector_cpu PRIVATE src/detect_obj.hpp src/detect_obj_cpu.cpp)
+target_sources(detector_cpu PRIVATE src/detect_obj.hpp src/detect_obj_cpu.cpp src/struct_utils.cpp)
 
 
 ##### GPU
