+++ src/blur_cpu.cpp
@@ -6,36 +6,6 @@
 #include <iostream>
 #include <cmath>
 
-// (1 / 2*pi*sigma^2) / e(-(x^2 + y^2)/2 * sigma^2)
-struct GaussianKernel* create_gaussian_kernel(unsigned char size) {
-    double **kernel = create2Dmatrix<double>(size, size);
-
-    int margin = (int) size / 2;
-    double sigma = 1.0;
-    double s = 2.0 * sigma * sigma;
-    
-    // sum is for normalization
-    double sum = 0.0;
-
-    for (int row = -margin; row <= margin; row++) {
-        for (int col = -margin; col <= margin; col++) {
-            const double radius = col * col + row * row;
-            kernel[row + margin][col + margin] = (exp(-radius / s)) / (M_PI * s);
-            sum += kernel[row + margin][col + margin];
-        }
-    }
-
-    // normalising the Kernel
-    for (unsigned char i = 0; i < size; ++i)
-        for (unsigned char j = 0; j < size; ++j)
-            kernel[i][j] /= sum;
-
-    struct GaussianKernel *g_kernel = (struct GaussianKernel *)std::malloc(sizeof(struct GaussianKernel));
-    g_kernel->kernel = kernel;
-    g_kernel->size = size;
-
-    return g_kernel;
-}
 
 unsigned char convolution(int x, int y, struct ImageMat *image, struct GaussianKernel* kernel)
 {
