+++ src/detect_obj.hpp
@@ -13,14 +13,30 @@ struct GaussianKernel {
     unsigned char size;
 };
 
+struct MorphologicalKernel {
+    unsigned char **kernel;
+    int size;
+};
+
+///// FILE: struct_utils.cpp
+struct ImageMat *new_matrix(int height, int width);
+
+///// FILE: struct_utils.cpp
 void swap_matrix(struct ImageMat *a, struct ImageMat *b);
 
+///// FILE: struct_utils.cpp
+struct GaussianKernel* create_gaussian_kernel(unsigned char size);
+
+///// FILE: struct_utils.cpp
+struct MorphologicalKernel* circular_kernel(int kernel_size);
+
+
 // FILE: detect_obj_cpu.cpp
-/// \param buffer_ref The RGBA24 image buffer
-/// \param buffer_obj The RGBA24 image buffer
-/// \param width Image width
-/// \param height Image height
-/// \param channels Image number of channels
+/// \param buffer_ref: The RGBA24 image buffer
+/// \param buffer_obj: The RGBA24 image buffer
+/// \param width: Image width
+/// \param height: Image height
+/// \param channels: Image number of channels
 unsigned char** detect_cpu(unsigned char* buffer_ref,
                            unsigned char* buffer_obj,
                            int width,
@@ -28,61 +44,51 @@ unsigned char** detect_cpu(unsigned char* buffer_ref,
                            int channels);
 
 // FILE: detect_obj_cpu.cpp
-/// \param buffer The RGBA24 image buffer
-/// \param width Image width
-/// \param height Image height
-/// \param channels Image number of channels
+/// \param src: The RGBA24 image buffer
+/// \param dst: The struct image to store the information
+/// \param width: Image width
+/// \param height: Image height
+/// \param channels: Image number of channels
 void to_gray_scale(unsigned char* src,
                    struct ImageMat* dst,
                    int width,
                    int height,
                    int channels);
 
-struct GaussianKernel* create_gaussian_kernel(unsigned char size);
 
 // FILE: detect_obj_cpu.cpp
-/// \param gray_ref The gray scale image buffer
-/// \param gray_obj The gray scale image buffer
-/// \param width Image width
-/// \param height Image height
+/// Store the result in the struct image obj
+/// \param ref: The struct image of the reference
+/// \param obj: The struct image of the object
 void difference(struct ImageMat *ref, struct ImageMat *obj);
 
 
 // FILE: blur_cpu.cpp
 // Apply gaussian blur to the image
-/// \param image The gray scale image
-/// \param width Image width
-/// \param height Image height
-/// \param kernel_size Kernel size
+/// \param image: The struct image
+/// \param temp: A temp struct image
+/// \param kernel: The Gaussian kernel 
 void apply_blurring(struct ImageMat* image,
                     struct ImageMat* temp,
                     struct GaussianKernel* kernel);
 
-void perform_dilation(unsigned char** input,
-                      unsigned char** temp,
-                      unsigned char** kernel,
-                      size_t height,
-                      size_t width,
-                      size_t kernel_size);
 
-void perform_erosion(unsigned char** input,
-                     unsigned char** temp,
-                     unsigned char** kernel,
-                     size_t height,
-                     size_t width,
-                     size_t kernel_size);
+// FILE: opening.cpp
+/// \param input: The struct image
+/// \param temp: A temp struct image
+/// \param kernel: A morphological kernel
+void perform_dilation(struct ImageMat* input,
+                      struct ImageMat* temp,
+                      struct MorphologicalKernel* kernel);
 
-unsigned char** perform_opening(unsigned char** input,
-                                 unsigned char** kernel,
-                                 size_t height,
-                                 size_t width,
-                                 size_t kernel_size);
+// FILE: opening.cpp
+/// \param input: The struct image
+/// \param temp: A temp struct image
+/// \param kernel: A morphological kernel
+void perform_erosion(struct ImageMat* input,
+                     struct ImageMat* temp,
+                     struct MorphologicalKernel* kernel);
 
-unsigned char** perform_closing(unsigned char** input,
-                                unsigned char** kernel,
-                                size_t height,
-                                size_t width,
-                                size_t kernel_size);
 
 unsigned char** compute_threshold(unsigned char** image, int width, int height);
 
@@ -97,6 +103,3 @@ void detect_gpu(unsigned char* buffer_ref,
                 int height,
                 int channels);
 
-unsigned char** circular_kernel(int kernel_size);
-
-void print_mat(unsigned char** input, size_t height, size_t width);
